move calendar's config into calendar's folder
